Fix CKAN 2.11 and Python 3.10 compatibility
This commit resolves multiple compatibility issues to make ckanext-scheming
work with CKAN 2.11 and Python 3.10:

1. Added missing dependencies (pycountry, ckanapi) to setup.py
2. Updated pytest to version 7.x for Python 3.10 compatibility
3. Fixed jinja2 Markup import for jinja2 3.0+ compatibility
4. Fixed validator resolution to use unicode_safe function instead of str type
5. Created separate decorators.py module for scheming_validator decorator
6. Updated all schema files to use unicode_safe instead of unicode validator
7. Removed six.text_type from default validators in plugins.py
8. Fixed auto_create_valid_name validator to prevent infinite loops

Files modified:
- setup.py: Added pycountry and ckanapi dependencies
- test-requirements.txt: Updated pytest and pytest-cov versions
- ckanext/scheming/decorators.py: New file with scheming_validator decorator
- ckanext/scheming/validation.py: Import unicode_safe, use ast.literal_eval
- ckanext/scheming/plugins.py: Removed six.text_type from default validators
- ckanext/scheming/unaids_validators.py: Fixed auto_create_valid_name bug
- ckanext/scheming/tests/test_form_snippets.py: Fixed Markup import
- Multiple schema JSON files: Replaced unicode with unicode_safe
- PROGRESS.md: Detailed documentation of all fixes

Test results improved significantly with these changes.
