+++ ckanext/scheming/ckan_dataset.json
@@ -49,7 +49,7 @@
     {
       "field_name": "version",
       "label": "Version",
-      "validators": "ignore_missing unicode package_version_validator",
+      "validators": "ignore_missing unicode_safe package_version_validator",
       "form_placeholder": "1.0"
     },
     {
