+++ ckanext/scheming/decorators.py
@@ -1,0 +1,9 @@
+def scheming_validator(fn):
+    """
+    Decorate a validator that needs to have the scheming fields
+    passed with this function. When generating navl validator lists
+    the function decorated will be called passing the field
+    and complete schema to produce the actual validator for each field.
+    """
+    fn.is_a_scheming_validator = True
+    return fn
