+++ ckanext/scheming/tests/schemas/auto_unique_validator.json
@@ -19,7 +19,7 @@
             "field_name": "name",
             "label": "URL",
             "form_snippet": "text.html",
-            "validators": "autogenerate auto_create_valid_name not_empty unicode name_validator package_name_validator",
+            "validators": "autogenerate auto_create_valid_name not_empty unicode_safe name_validator package_name_validator",
             "classes": ["hidden"],
             "template": "{}-autogenerate-{}",
             "template_args": ["location", "year"],
