+++ test-requirements.txt
@@ -1,6 +1,6 @@
 factory-boy>=2
 mock
 beautifulsoup4==4.8.2
-pytest==4.6.5
+pytest>=7.0.0
 pytest-ckan
-pytest-cov==2.7.1
+pytest-cov>=3.0.0
